[getting-started-NextJs] #9. Streaming
